Pull request 14: Upgrade dns-libs to 2.0
Merge in ADGUARD-CORE-LIBS/vpn-libs from feature/AG-16087 to master

Squashed commit of the following:

commit 686516bd213426f150db7895c5edffd40f134d7c
Author: Sergei Gunchenko <s.gunchenko@adguard.com>
Date:   Thu Sep 1 15:04:59 2022 +0300

    minor

commit a0a7641bb328ecdc5f061089fedf26c79f77f008
Author: Sergei Gunchenko <s.gunchenko@adguard.com>
Date:   Thu Sep 1 14:43:39 2022 +0300

    Upgrade dns-libs dep

commit 1d83f0794c5cbd8fc327e7f9a8a6a5220825c27c
Author: Sergei Gunchenko <s.gunchenko@adguard.com>
Date:   Thu Sep 1 14:42:47 2022 +0300

    Revert "try to let deps choose cppstd theirself"

    This reverts commit d87e2221814b8cee523f86bc64e2819cd73a7bf9.

commit d87e2221814b8cee523f86bc64e2819cd73a7bf9
Author: Sergei Gunchenko <s.gunchenko@adguard.com>
Date:   Thu Sep 1 14:08:06 2022 +0300

    try to let deps choose cppstd theirself

commit c1795afb6f3da3294b294625c4e866dcfd506052
Author: Sergei Gunchenko <s.gunchenko@adguard.com>
Date:   Wed Aug 31 14:37:47 2022 +0300

    suppress linter even more

commit 5a34674a0a8ab2effb6101156e2df3b726faa3ea
Author: Sergei Gunchenko <s.gunchenko@adguard.com>
Date:   Wed Aug 31 14:03:58 2022 +0300

    review

commit c6a5413615b996cc7a4bac0700790b3ac28106eb
Author: Sergei Gunchenko <s.gunchenko@adguard.com>
Date:   Wed Aug 31 13:47:16 2022 +0300

    Upgrade dns-libs to 2.0.2

commit 47f0230c93cf0933e79ec9e76921b630d1d45366
Author: Sergei Gunchenko <s.gunchenko@adguard.com>
Date:   Wed Aug 31 13:32:06 2022 +0300

    fix docker image version

commit b1f166501494bfb4cad5a304f5bfc39a1524349c
Author: Sergei Gunchenko <s.gunchenko@adguard.com>
Date:   Wed Aug 31 13:23:11 2022 +0300

    fix

commit 92f581a2f08ad31a08649b8b68dce9f2a1aced07
Author: Sergei Gunchenko <s.gunchenko@adguard.com>
Date:   Wed Aug 31 12:36:12 2022 +0300

    fix

commit 86e33576a1c9f0cd1ed62ca674c6cbca7dc89f2d
Author: Sergei Gunchenko <s.gunchenko@adguard.com>
Date:   Wed Aug 31 12:02:51 2022 +0300

    partial fix revert

commit cd03e67c017d8dfc864d3ac1fa77e3b42c1aa776
Author: Sergei Gunchenko <s.gunchenko@adguard.com>
Date:   Wed Aug 31 10:40:52 2022 +0300

    fix

commit 1969c867bd9f6426755b8bca515a3c80075504d5
Author: Sergei Gunchenko <s.gunchenko@adguard.com>
Date:   Tue Aug 30 17:31:51 2022 +0300

    fix

commit 875f0ad8faa06f77148029f2b7955d2949357923
Author: Sergei Gunchenko <s.gunchenko@adguard.com>
Date:   Tue Aug 30 16:35:23 2022 +0300

    fix

commit b126bfc1554146e1ed445c9a5ee8a2c64469127c
Author: Sergei Gunchenko <s.gunchenko@adguard.com>
Date:   Tue Aug 30 15:56:42 2022 +0300

    fix

commit a00dccef70b09fbb59b15309071f0c9c98f646b0
Author: Sergei Gunchenko <s.gunchenko@adguard.com>
Date:   Tue Aug 30 15:50:17 2022 +0300

    fix

commit 82ce653c9da1821c45abf7e3fdae1a21c6cd2127
Author: Sergei Gunchenko <s.gunchenko@adguard.com>
Date:   Tue Aug 30 15:09:19 2022 +0300

    remove ranges as they are still poorly supported

commit 00ad80212cda9809d38dbb257b47ca04cfd9231e
Author: Sergei Gunchenko <s.gunchenko@adguard.com>
Date:   Tue Aug 30 14:59:57 2022 +0300

    fix

commit 6a9ecaa2aeef91bd2c986c7b4bad725a5f8dae70
Author: Sergei Gunchenko <s.gunchenko@adguard.com>
Date:   Tue Aug 30 14:43:16 2022 +0300

    fix

commit 11f69beff46a6e20d54e44173c2f9803ee80aa5d
Author: Sergei Gunchenko <s.gunchenko@adguard.com>
Date:   Tue Aug 30 14:39:52 2022 +0300

    fix win

... and 23 more commits

+++ core/src/icmp_manager.cpp
@@ -98,7 +98,7 @@ IcmpManagerMessageStatus IcmpManager::register_request(const IcmpEchoRequest &re
 
     if (m_timer == nullptr) {
         m_timer.reset(event_new(vpn_event_loop_get_base(m_parameters.ev_loop), -1, EV_PERSIST, timer_callback, this));
-        const timeval tv = ms_to_timeval(m_parameters.request_timeout.value().count() / 10);
+        const timeval tv = ms_to_timeval(uint32_t(m_parameters.request_timeout.value().count() / 10));
         event_add(m_timer.get(), &tv);
     }
